Fix: Added explicit cast

+++ senc/utils/Socket_impl.hpp
@@ -142,7 +142,7 @@ namespace senc::utils
 		if (count < 0)
 			throw SocketException("Failed to recieve", Socket::get_last_sock_err());
 		auto [ip, port] = IP::from_underlying_sock_addr(addr);
-		return { count, ip, port };
+		return { static_cast<std::size_t>(count), ip, port };
 	}
 
 	template <IPType IP>
